requirements.txt
# Research dependencies
numpy>=1.24
pandas>=2.0
scipy>=1.10
ccxt>=4.0
optuna>=3.0
dwave-neal>=0.6
pyarrow>=12.0
